+++ src/napari_live_recording/control/devices/interface.py
@@ -2,10 +2,11 @@ import numpy as np
 from abc import abstractmethod
 from typing import Union, Tuple
 from qtpy.QtCore import QObject
-from napari_live_recording.common import ROI
+from napari_live_recording.common import ROI, ColorType
 from typing import Dict, List, Any
 from dataclasses import dataclass, replace
 from abc import ABC
+
 
 # the camera interface and setting management is heavily inspired by the work of Xavier Casas Moreno in ImSwitch
 # reference work: https://github.com/kasasxav/ImSwitch/blob/master/imswitch/imcontrol/model/managers/detectors/DetectorManager.py
@@ -15,6 +16,7 @@ class Parameter(ABC):
     editable: bool
     """Wether the parameter is readonly or not.
     """
+
 
 @dataclass
 class NumberParameter(Parameter):
@@ -30,6 +32,7 @@ class NumberParameter(Parameter):
     """Upper and lower boundaries of the possible parameter's value.
     """
 
+
 @dataclass
 class ListParameter(Parameter):
     value: str
@@ -39,13 +42,19 @@ class ListParameter(Parameter):
     options: List[str]
     """List of possible options for the parameter.
     """
-    
+
 
 class ICamera(QObject):
-    def __init__(self, name: str, deviceID: Union[str, int], parameters: Dict[str, Any], sensorShape: ROI) -> None:
+    def __init__(
+        self,
+        name: str,
+        deviceID: Union[str, int],
+        parameters: Dict[str, Any],
+        sensorShape: ROI,
+    ) -> None:
         """Generic camera device interface. Each device is initialized with a series of parameters.
         Live and recording are handled using child thread workers.
-        
+
         Args:
             name (str): name of the camera device.
             deviceID (`Union[str, int]`): device ID.
@@ -59,38 +68,46 @@ class ICamera(QObject):
         self.parameters = parameters
         self._roiShape = sensorShape
         self._fullShape = sensorShape
-        self.frameBuffer = None
-    
+        self._colorType = ColorType.GRAYLEVEL
+        try:
+            self.settingsWidget = self.settingsWidget
+        except:
+            pass
+
+    @property
+    def colorType(self) -> ColorType:
+        return self._colorType
+
     @property
     def fullShape(self) -> ROI:
         return self._fullShape
-    
+
     @property
     def roiShape(self) -> ROI:
         return self._roiShape
-    
+
     @roiShape.setter
     def roiShape(self, newROI: ROI) -> None:
         self._roiShape = replace(newROI)
-    
+
     @abstractmethod
     def setAcquisitionStatus(self, started: bool) -> None:
         """Sets the current acquisition status of the camera device.
+        - True: acquisition started;
+        - False: acquisition stopped.
         """
         raise NotImplementedError()
 
     @abstractmethod
     def grabFrame(self) -> np.ndarray:
-        """Returns the latest captured frame as a numpy array.
-        """
+        """Returns the latest captured frame as a numpy array."""
         raise NotImplementedError()
-    
+
     @abstractmethod
     def changeROI(self, newROI: ROI) -> None:
-        """Changes the Region Of Interest of the sensor's device.
-        """
+        """Changes the Region Of Interest of the sensor's device."""
         raise NotImplementedError()
-    
+
     @abstractmethod
     def changeParameter(name: str, value: Any) -> None:
         """Changes one of the settings of the device.
@@ -102,6 +119,11 @@ class ICamera(QObject):
         pass
 
     def close(self) -> None:
-        """Optional method to close the device.
-        """
+        """Optional method to close the device."""
         pass
+
+    def __enter__(self):
+        self.setAcquisitionStatus(True)
+
+    def __exit__(self, exc_type, exc_value, tb):
+        self.setAcquisitionStatus(False)